+++ src/CreateRecordsWrapper.js
@@ -1,27 +1,74 @@
-import React, { useState, useCallback } from 'react';
+import React, { useCallback } from 'react';
+import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { Button } from '@folio/stripes/components';
+import { stripesConnect } from '@folio/stripes/core';
 
-import CreateRecordsModal from './CreateRecordsModal';
+import CreateRecordsForm from './CreateRecordsForm';
+import { parseInstance } from './util';
+import {
+  useData,
+  useCallout,
+  useIsLoading,
+} from './hooks';
 
-const CreateRecordsWrapper = () => {
-  const [isModalOpen, toggleModal] = useState(false);
-  const openModal = useCallback(() => toggleModal(true), []);
-  const closeModal = useCallback(() => toggleModal(false), []);
+const CreateRecordsWrapper = ({
+  onCreate,
+  mutator: { createInstance },
+}) => {
+  const { identifierTypesByName } = useData();
+  const callout = useCallout();
+  const isLoading = useIsLoading();
+
+  const handleSubmit = useCallback(async (formData) => {
+    const { instance } = formData;
+
+    try {
+      await createInstance.POST(parseInstance(instance, identifierTypesByName));
+
+      callout.sendCallout({
+        message: <FormattedMessage id="ui-plugin-create-inventory-records.onSave.success" />,
+      });
+
+      onCreate();
+    } catch (error) {
+      callout.sendCallout({
+        message: <FormattedMessage id="ui-plugin-create-inventory-records.onSave.error" />,
+        type: 'error',
+      });
+    }
+  }, [onCreate, callout, createInstance, identifierTypesByName]);
+
+  if (isLoading) return null;
 
   return (
-    <>
-      <Button
-        data-test-add-inventory-records
-        marginBottom0
-        onClick={openModal}
-      >
-        <FormattedMessage id="ui-plugin-create-inventory-records.fastAddLabel" />
-      </Button>
-      {isModalOpen && <CreateRecordsModal onClose={closeModal} />}
-    </> //
+    <CreateRecordsForm
+      onSubmit={handleSubmit}
+      initialValues={{
+        instance: {
+          discoverySuppress: true,
+          contributors: [],
+        },
+        holding: {},
+        item: {},
+      }}
+    />
   );
 };
 
-export default CreateRecordsWrapper;
+CreateRecordsWrapper.propTypes = {
+  onCreate: PropTypes.func.isRequired,
+  mutator: PropTypes.object.isRequired,
+};
+
+CreateRecordsWrapper.manifest = Object.freeze({
+  createInstance: {
+    type: 'okapi',
+    records: 'instances',
+    throwErrors: false,
+    path: 'inventory/instances',
+    fetch: false,
+  },
+});
+
+export default stripesConnect(CreateRecordsWrapper);