+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as InstanceAccordion } from './InstanceAccordion';
+export { default as PrimaryToggleButton } from './PrimaryToggleButton';
+export { default as ContributorFields } from './ContributorFields';
+