+++ test/bigtest/network/models/holding.js
@@ -1,0 +1,6 @@
+import { Model, belongsTo, hasMany } from 'miragejs';
+
+export default Model.extend({
+  instance: belongsTo(),
+  items: hasMany(),
+});