+++ test/bigtest/network/models/instance.js
@@ -1,0 +1,6 @@
+import { Model, belongsTo, hasMany } from 'miragejs';
+
+export default Model.extend({
+  instanceType: belongsTo('instance-type'),
+  holdings: hasMany(),
+});