+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change history for ui-plugin-create-inventory-records
 
 ## 1.0.0 (IN PROGRESS)
+
+* Create instance via "fast add" form. Fixes UIPCIR-1.